backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain + Groq
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0

# LangGraph
langgraph>=0.2.0

# MCP
mcp>=1.0.0

# HTTP
httpx>=0.27.0

# Tavily (optional web search)
tavily-python>=0.3.0
